third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-audio-element/audio-loading-load-preload-metadata-deferred.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Audio with loading=lazy and preload=metadata that is not visible in viewport does not load audio data
  assert_equals: expected 0 but got 4
Harness: the test ran to completion.
